[REFACTOR] Moved the converter to respective weather service module
