+++ app/src/main/java/dev/hossain/weatheralert/ui/alertslist/CurrentAlertScreen.kt
@@ -82,9 +82,9 @@ import dagger.assisted.Assisted
 import dagger.assisted.AssistedFactory
 import dagger.assisted.AssistedInject
 import dev.hossain.weatheralert.data.AlertTileData
-import dev.hossain.weatheralert.data.CUMULATIVE_DATA_HOURS_24
 import dev.hossain.weatheralert.data.WeatherRepository
 import dev.hossain.weatheralert.data.icon
+import dev.hossain.weatheralert.datamodel.CUMULATIVE_DATA_HOURS_24
 import dev.hossain.weatheralert.datamodel.WeatherAlertCategory
 import dev.hossain.weatheralert.db.AlertDao
 import dev.hossain.weatheralert.di.AppScope
