+++ data-model/src/main/java/dev/hossain/weatheralert/datamodel/WeatherApiServiceResponse.kt
@@ -1,7 +1,8 @@
 package dev.hossain.weatheralert.datamodel
 
 /**
- * Marker interface for all weather API service response that must provide converted [AppForecastData].
+ * Marker interface for all weather API service response that must
+ * provide converted [AppForecastData] using [convertToForecastData].
  */
 interface WeatherApiServiceResponse {
     fun convertToForecastData(): AppForecastData
