+++ data-model/src/main/java/dev/hossain/weatheralert/datamodel/WeatherConstants.kt
@@ -1,0 +1,6 @@
+package dev.hossain.weatheralert.datamodel
+
+/**
+ * Total hours used to calculate the cumulative data for each type of weather alert.
+ */
+const val CUMULATIVE_DATA_HOURS_24 = 24
