+++ service/tomorrowio/src/main/java/io/tomorrow/api/model/TomorrowIo.kt
@@ -2,6 +2,8 @@ package io.tomorrow.api.model
 
 import com.squareup.moshi.Json
 import com.squareup.moshi.JsonClass
+import dev.hossain.weatheralert.datamodel.AppForecastData
+import dev.hossain.weatheralert.datamodel.WeatherApiServiceResponse
 
 /**
  * Represents the main response from the Tomorrow.io API.
@@ -12,7 +14,9 @@ data class WeatherResponse(
     @Json(name = "location") val location: Location,
     /** Timelines providing weather data across different intervals. */
     @Json(name = "timelines") val timelines: Timelines,
-)
+) : WeatherApiServiceResponse {
+    override fun convertToForecastData(): AppForecastData = this.toForecastData()
+}
 
 /**
  * Represents the real-time weather response from the Tomorrow.io API.
